+++ ios/Classes/ImeiPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface ImeiPlugin : NSObject<FlutterPlugin>
-@end